Fix false positive test_reference_not_found.
This test save empty Post object while it defines a title (which is
CharField) as pk. That forces to store post object with _id as ObjectId
value not a string. So post.delete() does not work and does not delete
any database objects as the post's pk is string not an ObjectId.

Test is still pass as comment.post reference field value is a string too.

To fix that while saving Post a title should be specified.