Aggregation add to queryset with sort
Fix for `pymongo.errors.OperationFailure: the $sort key specification must be an object`. This occurs when a queryset has been given an `order_by` then an aggregation is added. `self._order_by` is a list of tuples, but must be an object. The order of keys in `order_by` matters, so we must use `SON`.