Fix: Aggregation add to queryset with sort (#63)
Fix for `pymongo.errors.OperationFailure: the $sort key specification must be an object`. This occurs when a queryset has been given an `order_by` then an aggregation is added. `self._order_by` is a list of tuples, but must be an object. The order of keys in `order_by` matters, so we must use `SON`.

(cherry picked from commit b1186a8ec66cb6892e6d9c8052d9237eaf6aec43)